Move types into subfolder of backend and update imports across repo
